fix(android): BorrowError crash when loading assets on Android (#1505)
* fix(android): BorrowError crash when loading assets on Android

There is a race condition on the Android asset loading where the custom asset loading function is checked by the Java side while Rust is still setting it up, leading to a `already mutably borrowed: BorrowError` panic in the withAssetLoader function.
Looks like moving the function setup to happen before triggering the webview creation fixes it.

* use mutex